feat: expose rules
